requirements.txt
aiohttp==2.3.7
baidu-aip==2.0.0.1
beautifulsoup4==4.6.0
itchat==1.3.10
jieba==0.39
Pillow==5.0.0
requests==2.18.4